corrected the dtype in iwt

+++ fastmri_recon/models/subclassed_models/denoisers/mwcnn.py
@@ -79,7 +79,7 @@ class IWT(Layer):
         width = in_shape[2]
         # the number of channels can't be unknown for the convolutions
         n_channels = inputs.shape[3] // 4
-        outputs = tf.zeros([batch_size, 2 * height, 2 * width, n_channels])
+        outputs = tf.zeros([batch_size, 2 * height, 2 * width, n_channels], dtype=inputs.dtype)
         # for now we only consider greyscale
         x1 = inputs[..., 0:n_channels] / 2
         x2 = inputs[..., n_channels:2*n_channels] / 2
